Updated to VPX R3699

+++ RenderDevice.cpp
@@ -422,7 +422,7 @@ BOOL CALLBACK MonitorEnumList(__in  HMONITOR hMonitor, __in  HDC hdcMonitor, __i
    config.width = info.rcMonitor.right - info.rcMonitor.left;
    config.height = info.rcMonitor.bottom - info.rcMonitor.top;
    config.isPrimary = (config.top == 0) && (config.left == 0);
-   config.display = data->size(); // This number does neither map to the number form display settings nor something else.
+   config.display = (int)data->size(); // This number does neither map to the number form display settings nor something else.
 #ifdef ENABLE_SDL
    config.adapter = config.display;
 #else
